Separate rename edit construction
Amp-Thread-ID: https://ampcode.com/threads/T-019fa714-0bad-72ff-8885-6a565f7d4c13
